DE2695, DE2863, DE2832, DE704

+++ Discover-Bank-Phone/src/com/discover/mobile/bank/account/BankListAdapter.java
@@ -98,7 +98,7 @@ public class BankListAdapter extends ArrayAdapter<List<ActivityDetail>>{
 		/**Update the display values*/
 		holder.date.setText(convertDate(detail.dates.get(ActivityDetail.POSTED).split(ActivityDetail.DATE_DIVIDER)[0]));
 		holder.desc.setText(detail.description);
-		final double amount = detail.amount.value/DOLLAR_CONVERSION;
+		final double amount = ((double)detail.amount.value)/DOLLAR_CONVERSION;
 		if(amount < 0){
 			holder.amount.setText("-"+NumberFormat.getCurrencyInstance(Locale.US).format(amount*-1));
 		}else{
